# apps/server/openapi3.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/route-of-administration": {
      "get": {
        "operationId": "list-routes-of-administration",
        "summary": "List Routes of Administration",
        "description": "Returns all Route of Administrations.",
        "parameters": [
          {
            "name": "include",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route of Administrations found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteOfAdministrationModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "neuronek-api",
    "description": "Neuronek is a service that provides a RESTful API for providing substance information.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "name": "Bearer Token",
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer <token>\""
      }
    },
    "schemas": {
...
      "RouteOfAdministrationModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the route of administration",
            "example": "clvdzrfzj0000f2ftr6cm3fjr"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}